+++ example/lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:google_fonts/google_fonts.dart';
 import 'package:stylable_calendar/stylable_calendar.dart';
 
 void main() => runApp(MyApp());
@@ -28,11 +29,25 @@ class _HomePageState extends State<HomePage> {
       body: ListView(
         children: <Widget>[
           StylableCalendar(
-//            specialDays: [20, 13, 6, 10],
-//            highlightedDays: [1, 4, 5],
+            specialDays: [20, 13, 6, 10],
+            highlightedDays: [1, 4, 5],
 //            primaryColor: Theme.of(context).primaryColor,
 //            primaryColorDark: Theme.of(context).primaryColorDark,
 //            secondaryColor: Theme.of(context).primaryColorLight,
+//            headerTextStyle: GoogleFonts.pTSans(
+//              textStyle: TextStyle(
+//                color: Colors.red,
+//                fontSize: 18,
+//              ),
+//            ),
+//            dayTextStyle: TextStyle(
+//              color: Colors.lightGreen,
+//              fontSize: 18,
+//            ),
+//            dayNameTextStyle: TextStyle(
+//              fontSize: 16,
+//              color: Colors.orangeAccent,
+//            ),
             // in this case to animate highlighted dots
             selectedDate: (DateTime date) {
               print(date);